Move graphics::DrawParam to its own file
